[web] Embedded vtk-js vis works.
