+++ web/README.md
@@ -1,4 +1,30 @@
+## Start servers
+
+    hugo server
+
+If you want to modify css or javascript run `gulp` in another terminal:
+
+    gulp
+
 ## Used components
 
 - [Hugo](https://gothugo.com)
 - [flexboxgrid](http://flexboxgrid.com/)
+
+
+## Dump
+
+### Serve converted meshes with S3
+
+- Set CORS: http://stackoverflow.com/a/19557452/80480
+- Upload to e.g. http://s3-eu-central-1.amazonaws.com/jenkins-dump
+- Eventually add `crossdomain.xml` to the root of the bucket:
+
+```xml
+<?xml version="1.0"?>
+<!DOCTYPE cross-domain-policy SYSTEM
+"http://www.macromedia.com/xml/dtds/cross-domain-policy.dtd">
+<cross-domain-policy>
+  <allow-access-from domain="*" secure="false" />
+</cross-domain-policy>
+```
