+++ web/layouts/shortcodes/vis.html
@@ -1,0 +1,6 @@
+<iframe
+	src="./../../vis/index.html?fileURL=http://s3-eu-central-1.amazonaws.com/jenkins-dump/{{ .Get "url" }}"
+	width="100%"
+	{{ with .Get "height"}} height="{{.}}px"{{ else }} height="500px"{{ end }}
+>
+</iframe>
