+++ web/themes/ogs/layouts/partials/footer.html
@@ -1,6 +1,5 @@
 <!-- <script src="js/scripts.js"></script> -->
     </div> <!-- row -->
   </div> <!-- container -->
-<script src="/js/bundle.js"></script>
 </body>
 </html>
